pyproject.toml
[project]
name = "full-stack-database-management-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0",
    "django-admin-rangefilter>=0.13.5",
    "django-cors-headers>=4.9.0",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "streamlit>=1.52.1",
]
